# Optimise lowWatermark in Isolation