--- a/tsdb/isolation_test.go
+++ b/tsdb/isolation_test.go
@@ -14,52 +14,38 @@
 package tsdb
 
 import (
+	"strconv"
 	"sync"
 	"testing"
 )
 
-func TestIsolation(t *testing.T) {
-}
+func BenchmarkIsolation(b *testing.B) {
+	for _, goroutines := range []int{10, 100, 1000, 10000} {
+		b.Run(strconv.Itoa(goroutines), func(b *testing.B) {
+			iso := newIsolation()
 
-func BenchmarkIsolation_10(b *testing.B) {
-	benchmarkIsolation(b, 10)
-}
+			wg := sync.WaitGroup{}
+			start := make(chan struct{})
 
-func BenchmarkIsolation_100(b *testing.B) {
-	benchmarkIsolation(b, 100)
-}
+			for g := 0; g < goroutines; g++ {
+				wg.Add(1)
 
-func BenchmarkIsolation_1000(b *testing.B) {
-	benchmarkIsolation(b, 1000)
-}
+				go func() {
+					defer wg.Done()
+					<-start
 
-func BenchmarkIsolation_10000(b *testing.B) {
-	benchmarkIsolation(b, 10000)
-}
+					for i := 0; i < b.N; i++ {
+						appendID := iso.newAppendID()
+						_ = iso.lowWatermark()
 
-func benchmarkIsolation(b *testing.B, goroutines int) {
-	iso := newIsolation()
-
-	wg := sync.WaitGroup{}
-	start := make(chan struct{})
-
-	for g := 0; g < goroutines; g++ {
-		wg.Add(1)
-
-		go func() {
-			defer wg.Done()
-			<-start
-
-			for i := 0; i < b.N; i++ {
-				appendID := iso.newAppendID()
-				_ = iso.lowWatermark()
-
-				iso.closeAppend(appendID)
+						iso.closeAppend(appendID)
+					}
+				}()
 			}
-		}()
-	}
 
-	b.ResetTimer()
-	close(start)
-	wg.Wait()
+			b.ResetTimer()
+			close(start)
+			wg.Wait()
+		})
+	}
 }