@pstibrany 2020-06-03
> Great ideas, thank you very much.
> 
> I'm wondering if the `RWMutex` for `readMtx` and `appendMtx` is actually helping, especially now that the low-watermark retrieval is faster. I had a bunch of counter-intuitive results with `RWMutex` in the past. Ideally, we would test all four combinations of those two mutexes.
> 
> The other thing to keep an eye on is memary consumption. Let's see what the benchmark says…

I benchmarked accessMtx with Mutex and RWMutex and there was noticeable improvement. I can rerun benchmarks for all 4 combinations and post results here.