@beorn7 2020-06-03
Yeah, that would be great. It's of course also the question how well the benchmarks model real-world usage. The automatic end-to-end benchmark suite we run now on this PR is probably the best we have for that, but not sure if it is worth the effort to run all four combinations with it.