@pstibrany 2020-06-03
| | `accessMtx RWLock` | `accessMtx Lock`| 
|-|-|-|
| `readMtx RWLock` | a.txt | d.txt |
| `readMtx Lock` | b.txt | c.txt |

Results:
```
name \ time/op              a.txt        b.txt        c.txt         d.txt
IsolationWithState/10-4     4.11µs ± 2%  3.97µs ± 1%   3.39µs ± 5%   3.37µs ± 8%
IsolationWithState/100-4    31.4µs ± 0%  31.4µs ± 1%   66.9µs ± 4%   67.9µs ± 4%
IsolationWithState/1000-4    456µs ± 2%   476µs ± 4%   1046µs ± 1%   1034µs ± 2%
IsolationWithState/10000-4  4.03ms ± 2%  4.13ms ± 3%  37.29ms ±13%  36.65ms ±25%

name \ alloc/op             a.txt        b.txt        c.txt         d.txt
IsolationWithState/10-4       177B ± 0%    178B ± 0%     181B ± 0%     182B ± 1%
IsolationWithState/100-4    1.09kB ± 1%  1.10kB ± 2%   1.57kB ± 1%   1.58kB ± 1%
IsolationWithState/1000-4    103kB ± 9%   104kB ± 2%    114kB ± 0%    114kB ± 0%
IsolationWithState/10000-4   329kB ±18%   371kB ±10%   8277kB ±15%   8076kB ±27%

name \ allocs/op            a.txt        b.txt        c.txt         d.txt
IsolationWithState/10-4       3.00 ± 0%    3.00 ± 0%     3.00 ± 0%     3.00 ± 0%
IsolationWithState/100-4      5.00 ± 0%    5.00 ± 0%     4.00 ± 0%     4.00 ± 0%
IsolationWithState/1000-4     96.8 ± 3%    96.4 ± 2%    120.2 ± 2%    121.6 ± 0%
IsolationWithState/10000-4     653 ± 5%     680 ± 5%     2923 ±18%     2966 ±49%
```

I will push `BenchmarkIsolationWithState` -- it's basically same as Isolation, except it also adds 1/100 of `goroutines` calling just `State()` and `Close()`.