@geekodour 2020-06-17
/funcbench master BenchmarkIsolation.*

Testing funcbench