@codesome 2020-06-17
`BenchmarkIsolation/10000-4` `BenchmarkIsolationWithState/10000-4` have a big differences. I would expect them to be similar as this PR is in master and no other code changes in this area. How are these run?