@krasi-georgiev 2020-06-17
yep I was thinking the same thing. 

@geekodour maybe in the initial comment should also post the full command that will be  run 

```
stopwatch Welcome to Funcbench Tool. stopwatch

Running benchmark BenchmarkIsolation.* on PR-7332 vs master
go test ....
```