@krasi-georgiev 2020-06-17
or actually maybe this whould be along the results as the comment monitor doesn't know upfront which command will be run