@krasi-georgiev 2020-06-29
/funcbench master github.com/prometheus/prometheus/promql