@krasi-georgiev 2020-06-29
@pstibrany ooh yeah you can ignore all these, we are testing a new micro benchmarking tool, sorry for the noise. We expected that there should be no change in the results but for some reason this running the same test results in slightly different results. Will continue these in a separate PR