@pstibrany 2020-06-29
Ah, thanks for letting me know :) I wasn’t sure if there was something wrong with the PR or not. Feel free to continue using it, I will unsubscribe from it.